Refs #9804 -- Fixed test for sequence reset of M2M with inherited through model.
